refactor: group header components
